% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata.R
\name{dataanalysis}
\alias{dataanalysis}
\title{Preprocessing of trading volume and price data}
\usage{
dataanalysis(p0)
}
\arguments{
\item{p0}{is a matrix with two columns,the first column is the trading volume of the stock,
while the second is price of the stock}
}
\description{
Preprocessing of trading volume and price data
}
\examples{
\dontrun{
a=rep(5,times=10)
b=abs(runif(10))
c=cbind(a,b)
dataanalysis(c)
}
}
