% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata.R
\name{OU.lik}
\alias{OU.lik}
\title{likelihood of parametric estimation for the stochastic diffenrential equation}
\usage{
OU.lik(theta1, theta2, theta3, factor)
}
\arguments{
\item{theta1}{is a parameter}

\item{theta2}{is a parameter}

\item{theta3}{is a parameter}

\item{factor}{is a vector of time series}
}
\description{
Parametric estimation for the Vasicek model using maximum likelihood ofEuler-Maruyama method .
The parametric form of the Vasicek model used here is given by
\deqn{dX_t = (\thata1 - \theta2 X_t)dt + \theta3 dW_t.}
}
\examples{
\dontrun{
  z=ts(abs(rnorm(100))
  OU.lik(0.5,-0.5,0.3,z)
}
}
