\name{SA23204198-package}
\alias{SA23204198-package}
\docType{package}
\title{This package provides several functions for anaylsis of time series data}
\description{R and Rcpp versions of gibbs sampler and aother function are provided}
\author{
Chengbo Zhang
Maintainer: Chengbo Zhang <3115632878@qq.com>
}
\keyword{ package }